refactor(rust): split commands.rs and lib.rs into focused modules
commands.rs was 676 lines covering file I/O, directory walking, and
wikilink scanning. Split into a commands/ submodule:

  - commands/file.rs       — read_file, write_file, get_file_metadata,
                              get_initial_file, print_document
  - commands/directory.rs  — read_directory, list_markdown_files
  - commands/wikilinks.rs  — scan_wikilinks + helpers
  - commands/walk.rs       — shared walker constants

Also extracted the markdown extension check (is_markdown_file +
MD_EXTENSIONS include) from lib.rs into a new markdown.rs, leaving
lib.rs as pure setup/wiring (124 lines, down from 206).

Pure code motion — tests travel with their code, no behavior change.
All 39 unit tests pass.
